Rewriting the Endomondo JSON parser with merge-then-bind on Jackson 3
Parsing extracted from UploadController into EndomondoJsonParser: the
single-key-map structure is merged into plain object nodes (recursively for
points and location) and bound to the DTOs. Fixes all pinned defects: real
export files parse end-to-end, points attach with both coordinates and full
metrics, decimal calories and integer altitudes convert tolerantly
(calories_kcal and Point.speed_kmh are now Double, matching real files).
InvalidWorkoutJsonException is the parser's single failure mode; the upload
flow now handles wrong-shape JSON with the user-facing message instead of a
500. Jackson 2 pin dropped; characterization tests flipped to a regression
suite (EndomondoJsonParserRegressionTest) plus a MockMvc upload-flow test.
E2E verified: a real 2011 workout file uploads successfully. OpenSpec
change: fix-json-parser.

+++ src/main/java/pl/kiraga/endomondoexportparser/format/json/EndomondoJson.java
@@ -22,7 +22,7 @@ public class EndomondoJson {
     private String end_time;
     private Integer duration_s;
     private Double distance_km;
-    private Integer calories_kcal;
+    private Double calories_kcal;
     private Double altitude_min_m;
     private Double altitude_max_m;
     private Double speed_avg_kmh;
